Move renderers to manager
This is the first step towards allowing customisation of renderers.
