+++ jupyter-widgets-htmlmanager/src/output.ts
@@ -7,60 +7,11 @@ import { Panel, Widget } from '@phosphor/widgets';
 
 import { OutputAreaModel, OutputArea } from '@jupyterlab/outputarea';
 
-import { IRenderMime, RenderMime } from '@jupyterlab/rendermime';
-
 import { HTMLManager } from './htmlmanager';
 
 import * as $ from 'jquery';
 
 import '../css/output.css';
-
-const WIDGET_MIMETYPE = 'application/vnd.jupyter.widget-view+json';
-
-export class WidgetRenderer implements RenderMime.IRenderer {
-    constructor(widget_manager: HTMLManager) {
-        this._manager = widget_manager;
-    }
-
-    canRender(options: RenderMime.IRenderOptions): boolean {
-        const source: any = options.model.data.get(options.mimeType);
-        const model = this._manager.get_model(source.model_id);
-        return model !== void 0;
-    }
-
-    wouldSanitize(options: RenderMime.IRenderOptions): boolean {
-        return false;
-    }
-
-    render(options: RenderMime.IRenderOptions) {
-        const widget = new Panel();
-        const source: any = options.model.data.get(options.mimeType);
-        const model = this._manager.get_model(source.model_id);
-        console.log(model);
-        if (model) {
-            model.then((model: any) => {
-                return this._manager.display_model(null, model, {
-                    el: widget.node
-                })
-            }).catch((error) => {
-                console.log('Error creating widget.')
-                console.log(error);
-                const node = document.createElement('p')
-                node.textContent = 'Error creating widget.'
-                widget.addWidget(new Widget({node}))
-            })
-        } else {
-            const node = document.createElement('p')
-            node.textContent = 'Error creating widget.'
-            widget.addWidget(new Widget({node}))
-        }
-        return widget;
-    }
-
-    private _manager: HTMLManager;
-    public mimeTypes = [WIDGET_MIMETYPE];
-}
-
 
 export class OutputModel extends DOMWidgetModel {
     defaults() {
@@ -103,11 +54,8 @@ export class OutputView extends DOMWidgetView {
     }
 
     render() {
-        const rendermime = new RenderMime({items: RenderMime.getDefaultItems()})
-        rendermime.addRenderer({
-            mimeType: WIDGET_MIMETYPE,
-            renderer: new WidgetRenderer(this.model.widget_manager)
-        });
+        const manager = this.model.widget_manager
+        const rendermime = manager.renderMime;
         this._outputView = new OutputArea({
             rendermime: rendermime,
             contentFactory: OutputArea.defaultContentFactory,
